virtinst: Add support for fs write policy
This introduces new attribute wrpolicy with only supported
value as immediate. This helps specify whether to skip the
host page cache.

When wrpolicy is specified, meaning when wrpolicy=immediate
a writeback is explicitly initiated for the dirty pages in
the host page cache as part of the guest file write operation.

Signed-off-by: Deepak C Shetty <deepakcs@linux.vnet.ibm.com>