virtinst:Add spapr-vio address type for pseries kvm guest
For pseries kvm guest, scsi controller can use
spapr-vio address type, not only pci address type.
So it's necessary to provide such an option to
device to configure it.

This patch provides spapr-vio address type option for
scsi controller. And scsi controller and address type
tags are generated in xml file. It is like the following:

<controller type='scsi' index='0'>
    <address type='spapr-vio'/>
</controller>

Signed-off-by: Li Zhang <zhlcindy@linux.vnet.ibm.com>