Support multiple seclabels
Until now, virt-install supported only one seclabel and it was the
first one libvirt reported in capabilities.  This patch adds support
for more of them and also adds a functionality to try to match the
right one from the label given.  This is done by checking how many
colons the label has (precisely said, to how many parts it is split
by the colons).

(crobinso: add Martin to AUTHORS)