Add pre-generated virt-install.pod to fix freshly-cloned repos
Feeling really dull, I'm adding the 'man/en/virt-install.pod'
(generated on current HEAD), because I wrongly assumed that it is
being kept in the repo when redesigning the build with its
dependencies, but I missed that it was listed in '.gitignore'.

The file should be kept in the repository in order for the build to be
working in a freshly cloned repo.  The file then gets updated whenever
'python setup.py sdist' is ran, hence it is kept up-to date with
releases.