Don't support comments in parsed parameters
Until now, when virtinst was parsing parameters, '#' were used as
starts of a comment, '\' as escape strings etc. due to the use of
shlex.  The use of '\' is predictable (e.g. filename with a ',' in
it), but the '#' is usually escaped in shell only.  Also, even when
not escaped, shlex is not fully compatible with some syntax rules for
parsing '#'.

For example '#' in 'echo asdf#fdsa' will not be used as a comment in
bash (there is no whitespace in front of the '#'), but in our '--disk
name#something.img' (parsed using shlex), it will be used as a comment
even though it is not expected from the user.

Removing '#' (the only character) from shlex.commenters disables all
similar comments, which is desired as users won't be trying to use
comments to comment out parts of parameters.